requirements.txt
invoke
flake8
black
bandit
mypy
isort
jupyterlab
ipykernel
ipywidgets
python-dotenv
pypdf2
gradio
langchain
huggingface_hub
transformers
einops
xformers
tqdm
accelerate
fastapi
uvicorn
requests
types-requests